Dialogue, Choice + Other System Updates
Dialogue now has soft-wrapping.
Choice boxes are now gone in favour of text being centred and related dialogue to be visible.
Fixed the chapter system from being unable to load the next one depending on choice amount, leading to a game crash.
Gave the game the ability to return to Chapter 1 if the save file is missing.
If the choice data file is missing then it will reset to default values.
Added an empty __init__.py file in Game2 Chapters folder so that it would be counted as a "package".
When starting a New Game, choices.json PROPERLY resets and doesn't create it into the same folder where the launcher is located at.